+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # ChangeLog
 
-* Added set dynamic machine level.
+* Added supports of setting dynamic machine level.
+* Added supports of modifying geckolib animation speed.
 * Fixed In world preview rendering error.